requirements.txt
--index-url https://pypi.org/simple
--extra-index-url https://download.pytorch.org/whl/cpu

# Core RAG/Data Packages
torch==2.6.0+cpu
sentence-transformers==3.3.1
pypdf==5.1.0
numpy==1.26.4
pandas==2.3.3

# LangChain / LangGraph stack
langchain==0.3.7
langchain-openai>=0.2.8
langchain-pinecone>=0.2.0
langchain-community>=0.3.7
langchain-text-splitters>=0.3.2
langgraph>=0.2.39

# Vector store
pinecone[grpc]>=5.3.1

# API server
fastapi>=0.115.5
uvicorn[standard]>=0.32.1
python-multipart==0.0.19
python-dotenv>=1.0.1
pydantic>=2.9.2

# Testing
pytest
httpx
flake8